Refactor routing configuration and add new components
This commit refactors the routing configuration in the app-routing.module.ts file. It updates the import statements for the RouterModule and Routes modules. Additionally, it adds new routes for the HomePageComponent, SearchPageComponent, FavoritePageComponent, BookDetailsComponent, and AuthorDetailsComponent.

It also adds the necessary declarations and imports for the new components in the app.module.ts file.

Lastly, it adds the template files for the AuthorDetailsComponent, BookDetailsComponent, FavoritePageComponent, HomePageComponent, and SearchPageComponent.